requirement.txt
python-telegram-bot
fastapi
pydantic